fuzzing: refactored main-fuzzing files.
